@@ -1,7 +1,9 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:fluttertoast/fluttertoast.dart';
+import 'package:stuff_pages/enums/searchEnum.dart';
 import 'package:stuff_pages/request/entities/achievement.dart';
 import 'package:stuff_pages/request/entities/game.dart';
 import 'package:stuff_pages/request/http.dart';
@@ -11,7 +13,7 @@ import 'package:stuff_pages/utils/gameUtil.dart';
 import '../navigator.dart';
 
 class ShowTrophy extends StatefulWidget {
-  Game game;
+  late Game game;
 
   ShowTrophy(Game game) {
     this.game = game;
@@ -22,7 +24,7 @@ class ShowTrophy extends StatefulWidget {
 }
 
 class _ShowTrophyState extends State<ShowTrophy> {
-  Game game;
+  late Game game;
   bool donefilter = false;
   List<Achievement> _achievements = [];
   List<Achievement> filteredAchievments = [];
@@ -52,7 +54,7 @@ class _ShowTrophyState extends State<ShowTrophy> {
     filteredAchievments.clear();
     filteredAchievments.addAll(_achievements);
     _achievements.forEach((achievement) {
-      if (!achievement.earned == donefilter) {
+      if (!achievement.earned! == donefilter) {
         filteredAchievments.remove(achievement);
       }
     });
@@ -61,14 +63,15 @@ class _ShowTrophyState extends State<ShowTrophy> {
 
   _getTrophies() {
     filteredAchievments.clear();
-    final endpoint = "achievements/game=" + game.gameId;
+    final endpoint = "achievements/game=" + game.gameId!;
     Api.get(endpoint).then((res) {
       setState(() {
         Iterable list = json.decode(res.body);
         _achievements = list.map((e) => Achievement.fromJson(e)).toList();
-        _achievements.sort((a, b) => a.title.compareTo(b.title));
+        _achievements.sort((a, b) => a.title!.compareTo(b.title!));
         filteredAchievments.addAll(_achievements);
-        donefilter = game.earned != null && (game.earned / game.sum * 100) == 100;
+        donefilter =
+            game.earned != null && (game.earned! / game.sum! * 100) == 100;
         filter();
       });
     });
@@ -88,8 +91,8 @@ class _ShowTrophyState extends State<ShowTrophy> {
     return Scaffold(
       appBar: AppBar(
           backgroundColor: backgroundColor,
-          title: Text(game.title, style: TextStyle(color: fontColor)),
-          actions: <Widget>[doneFilter()]),
+          title: Text(game.title!, style: TextStyle(color: fontColor)),
+          actions: <Widget>[starButton(), doneFilter()]),
       body: Center(
         child: Column(
           children: <Widget>[Expanded(child: _trophyList())],
@@ -105,40 +108,24 @@ class _ShowTrophyState extends State<ShowTrophy> {
       itemCount: filteredAchievments.length,
       itemBuilder: (context, index) {
         final item = filteredAchievments[index];
-        return InkWell(
-          child: Dismissible(
+        return Slidable(
             key: UniqueKey(),
-            child: Card(
-              child: getTrophy(item),
-              color: cardBackgroundColor,
+            endActionPane: ActionPane(
+              motion: const ScrollMotion(),
+              children: [youtubeButton(game, item), googleButton(game, item)],
             ),
-            onDismissed: (direction) {
-              Fluttertoast.showToast(
-                  msg: item.title,
-                  toastLength: Toast.LENGTH_SHORT,
-                  gravity: ToastGravity.BOTTOM,
-                  timeInSecForIosWeb: 3,
-                  backgroundColor: item.earned ? deleteColor : addedColor,
-                  fontSize: 16.0);
-              setState(() {
-                item.earned = !item.earned;
-                Api.put('achievements/', item, item.id);
-                filter();
-              });
-            },
-            background:
-                Container(color: item.earned ? deleteColor : addedColor),
-          ),
-          onTap: () {
-            launchURL(game.title + " " + item.title);
-          },
-        );
+            child: InkWell(
+              child: Card(
+                child: getTrophy(item),
+                color: cardBackgroundColor,
+              ),
+            ));
       },
     );
   }
 
   Widget getTrophy(Achievement trophy) {
-    final secret = trophy.secret && !trophy.show;
+    final secret = trophy.secret! && !trophy.show!;
     final earned = trophy.earned;
     return Column(
       mainAxisSize: MainAxisSize.min,
@@ -152,37 +139,57 @@ class _ShowTrophyState extends State<ShowTrophy> {
                 maxHeight: 200,
               ),
               child: trophyImg(trophy)),
-          title: secret && !earned ? Text(secretTitle, style: TextStyle(color: fontColor)) : Text(trophy.title, style: TextStyle(color: fontColor)),
-          subtitle: secret && !earned
+          title: secret && !earned!
+              ? Text(secretTitle, style: TextStyle(color: fontColor))
+              : Text(trophy.title!, style: TextStyle(color: fontColor)),
+          subtitle: secret && !earned!
               ? Text(secretDescription, style: TextStyle(color: fontColor))
-              : Text(trophy.description, style: TextStyle(color: fontColor)),
+              : Text(trophy.description!, style: TextStyle(color: fontColor)),
           trailing: showButton(trophy),
-          onTap: () {
-            launchURL(game.title + " " + trophy.title);
+          onLongPress: () {
+            setState(() {
+              trophy.earned = !trophy.earned!;
+              Api.put('achievements/', trophy, trophy.id);
+              filter();
+            });
           },
         ),
       ],
     );
   }
 
-  Widget showButton(Achievement trophy) {
-    if (trophy.secret) {
+  Widget? showButton(Achievement trophy) {
+    if (trophy.secret! && !trophy.earned!) {
       return IconButton(
-          icon: trophy.show
+          icon: trophy.show!
               ? Icon(
-            Icons.lock_open_outlined,
-            color: fontColor,
-          )
+                  Icons.lock_open_outlined,
+                  color: fontColor,
+                )
               : Icon(
-            Icons.lock_outlined,
-            color: fontColor,
-          ),
+                  Icons.lock_outlined,
+                  color: fontColor,
+                ),
           onPressed: () {
             setState(() {
-              trophy.show = !trophy.show;
+              trophy.show = !trophy.show!;
             });
           });
     }
     return null;
+  }
+
+  Widget starButton() {
+    return IconButton(
+        icon: Icon(
+          game.star! ? Icons.star : Icons.star_border,
+          color: futureColor,
+        ),
+        onPressed: () {
+          setState(() {
+            game.star = !game.star!;
+            Api.put("games/", game, game.id);
+          });
+        });
   }
 }
