@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:fluttertoast/fluttertoast.dart';
+import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:stuff_pages/request/entities/achievement.dart';
 import 'package:stuff_pages/request/entities/game.dart';
 import 'package:stuff_pages/request/http.dart';
@@ -11,7 +11,7 @@ import 'package:stuff_pages/utils/gameUtil.dart';
 import '../navigator.dart';
 
 class ShowAchievement extends StatefulWidget {
-  Game game;
+  late Game game;
 
   ShowAchievement(Game game) {
     this.game = game;
@@ -22,7 +22,7 @@ class ShowAchievement extends StatefulWidget {
 }
 
 class _ShowAchievementState extends State<ShowAchievement> {
-  Game game;
+  Game game = new Game();
   bool donefilter = false;
   List<Achievement> _achievements = [];
   List<Achievement> filteredAchievments = [];
@@ -53,7 +53,7 @@ class _ShowAchievementState extends State<ShowAchievement> {
     filteredAchievments.clear();
     filteredAchievments.addAll(_achievements);
     _achievements.forEach((achievement) {
-      if (!achievement.earned == donefilter) {
+      if (!achievement.earned! == donefilter) {
         filteredAchievments.remove(achievement);
       }
     });
@@ -62,15 +62,15 @@ class _ShowAchievementState extends State<ShowAchievement> {
 
   _getAchievements() {
     filteredAchievments.clear();
-    final endpoint = "achievements/game=" + game.gameId;
+    final endpoint = "achievements/game=" + game.gameId!;
     Api.get(endpoint).then((res) {
       setState(() {
         Iterable list = json.decode(res.body);
         _achievements = list.map((e) => Achievement.fromJson(e)).toList();
-        _achievements.sort((a, b) => a.title.compareTo(b.title));
+        _achievements.sort((a, b) => a.title!.compareTo(b.title!));
         filteredAchievments.addAll(_achievements);
         donefilter =
-            game.earned != null && (game.earned / game.sum * 100) == 100;
+            game.earned != null && (game.earned! / game.sum! * 100) == 100;
         filter();
       });
     });
@@ -90,8 +90,8 @@ class _ShowAchievementState extends State<ShowAchievement> {
     return Scaffold(
       appBar: AppBar(
           backgroundColor: backgroundColor,
-          title: Text(game.title, style: TextStyle(color: fontColor)),
-          actions: <Widget>[refreshButton(), doneFilter()]),
+          title: Text(game.title!, style: TextStyle(color: fontColor)),
+          actions: <Widget>[starButton(), doneFilter()]),
       body: Center(
         child: Column(
           children: <Widget>[Expanded(child: _achievementList())],
@@ -107,40 +107,24 @@ class _ShowAchievementState extends State<ShowAchievement> {
       itemCount: filteredAchievments.length,
       itemBuilder: (context, index) {
         final item = filteredAchievments[index];
-        return InkWell(
-          child: Dismissible(
+        return Slidable(
             key: UniqueKey(),
-            child: Card(
-              child: getAchievement(item),
-              color: cardBackgroundColor,
+            endActionPane: ActionPane(
+              motion: const ScrollMotion(),
+              children: [youtubeButton(game, item), googleButton(game, item)],
             ),
-            onDismissed: (direction) {
-              Fluttertoast.showToast(
-                  msg: item.title,
-                  toastLength: Toast.LENGTH_SHORT,
-                  gravity: ToastGravity.BOTTOM,
-                  timeInSecForIosWeb: 3,
-                  backgroundColor: item.earned ? deleteColor : addedColor,
-                  fontSize: 16.0);
-              setState(() {
-                item.earned = !item.earned;
-                Api.put('achievements/', item, item.id);
-                filter();
-              });
-            },
-            background:
-                Container(color: item.earned ? deleteColor : addedColor),
-          ),
-          onTap: () {
-            launchURL(game.title + " " + item.title);
-          },
-        );
+            child: InkWell(
+              child: Card(
+                child: getAchievement(item),
+                color: cardBackgroundColor,
+              ),
+            ));
       },
     );
   }
 
   Widget getAchievement(Achievement achievement) {
-    final secret = achievement.secret && !achievement.show;
+    final secret = achievement.secret! && !achievement.show!;
     final earned = achievement.earned;
     return Column(
       mainAxisSize: MainAxisSize.min,
@@ -154,26 +138,30 @@ class _ShowAchievementState extends State<ShowAchievement> {
                 maxHeight: 200,
               ),
               child: achievementImg(achievement)),
-          title: secret && !earned
+          title: secret && !earned!
               ? Text(secretTitle, style: TextStyle(color: fontColor))
-              : Text(achievement.title, style: TextStyle(color: fontColor)),
-          subtitle: secret && !earned
+              : Text(achievement.title!, style: TextStyle(color: fontColor)),
+          subtitle: secret && !earned!
               ? Text(secretDescription, style: TextStyle(color: fontColor))
-              : Text(achievement.description,
+              : Text(achievement.description!,
                   style: TextStyle(color: fontColor)),
           trailing: showButton(achievement),
           onTap: () {
-            launchURL(game.title + " " + achievement.title);
+            setState(() {
+              achievement.earned = !achievement.earned!;
+              Api.put('achievements/', achievement, achievement.id);
+              filter();
+            });
           },
         ),
       ],
     );
   }
 
-  Widget showButton(Achievement achievement) {
-    if (achievement.secret) {
+  Widget? showButton(Achievement achievement) {
+    if (achievement.secret! && !achievement.earned!) {
       return IconButton(
-          icon: achievement.show
+          icon: achievement.show!
               ? Icon(
                   Icons.lock_open_outlined,
                   color: fontColor,
@@ -184,24 +172,23 @@ class _ShowAchievementState extends State<ShowAchievement> {
                 ),
           onPressed: () {
             setState(() {
-              achievement.show = !achievement.show;
+              achievement.show = !achievement.show!;
             });
           });
     }
     return null;
   }
 
-  Widget refreshButton() {
+  Widget starButton() {
     return IconButton(
         icon: Icon(
-          Icons.refresh,
+          game.star! ? Icons.star : Icons.star_border,
           color: futureColor,
         ),
         onPressed: () {
           setState(() {
-            String endpoint = 'achievements/game=' + this.game.gameId;
-            Api.post(endpoint, []);
-            initState();
+            game.star = !game.star!;
+            Api.put("games/", game, game.id);
           });
         });
   }
